refactor(mcp): eliminate shims, use rolldown config and Nx-native build
- Extract IAC_PROVIDERS to utils/iac-providers.ts (no @nx/devkit dep)
- Move listGenerators and NxGeneratorInfo to utils/generators.ts
- Update MCP server imports to use new lightweight util modules
- Keep re-exports in utils/nx.ts and utils/iac.ts for backward compat
- Replace build.mjs with rolldown.config.mjs + project.json commands
- Remove shim files (no longer needed)
- Add shebang to entry point source (rolldown preserves it)
