+++ packages/nx-plugin-mcp/project.json
@@ -6,14 +6,38 @@
   "tags": [],
   "targets": {
     "compile": {
+      "cache": true,
       "executor": "nx:run-commands",
       "outputs": ["{workspaceRoot}/dist/packages/nx-plugin-mcp"],
       "options": {
-        "command": "node packages/nx-plugin-mcp/build.mjs"
+        "command": "rolldown -c rolldown.config.mjs",
+        "cwd": "{projectRoot}"
       }
     },
-    "build": {
+    "post-compile": {
+      "executor": "nx:run-commands",
+      "options": {
+        "commands": ["chmod +x dist/packages/nx-plugin-mcp/bin/aws-nx-mcp.cjs"]
+      },
       "dependsOn": ["compile"]
+    },
+    "copy-assets": {
+      "executor": "nx:run-commands",
+      "outputs": ["{workspaceRoot}/dist/packages/nx-plugin-mcp"],
+      "options": {
+        "commands": [
+          "cp packages/nx-plugin-mcp/package.json dist/packages/nx-plugin-mcp/package.json",
+          "cp packages/nx-plugin-mcp/README.md dist/packages/nx-plugin-mcp/README.md",
+          "test -f packages/nx-plugin-mcp/LICENSE && cp packages/nx-plugin-mcp/LICENSE dist/packages/nx-plugin-mcp/LICENSE || true",
+          "cp packages/nx-plugin/generators.json dist/packages/nx-plugin-mcp/generators.json",
+          "cd packages/nx-plugin && find src -name 'schema.json' -exec sh -c 'mkdir -p ../../dist/packages/nx-plugin-mcp/$(dirname $1) && cp $1 ../../dist/packages/nx-plugin-mcp/$1' _ {} \\;"
+        ],
+        "parallel": false
+      },
+      "dependsOn": ["compile"]
+    },
+    "build": {
+      "dependsOn": ["compile", "post-compile", "copy-assets"]
     }
   }
 }
