+++ packages/nx-plugin-mcp/src/index.ts
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 /**
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
