--- packages/nx-plugin-mcp/src/shims/nx.ts
@@ -1,6 +1,0 @@
-/**
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-export type { GeneratorInfo as NxGeneratorInfo } from '../../../nx-plugin/src/utils/generators';
-export { buildGeneratorInfoList } from '../../../nx-plugin/src/utils/generators';
