+++ packages/nx-plugin/src/mcp-server/tools/general-guidance.ts
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
-import { NxGeneratorInfo } from '../../utils/nx';
+import { NxGeneratorInfo } from '../../utils/generators';
 import { PACKAGE_MANAGERS } from '../schema';
 import { buildNxCommand, fetchGuidePages } from '../generator-info';
-import { IAC_PROVIDERS } from '../../utils/iac';
+import { IAC_PROVIDERS } from '../../utils/iac-providers';
 
 export const TOOL_SELECTION_GUIDE = `## Tool Selection Guide
