+++ packages/nx-plugin/src/utils/nx.ts
@@ -13,29 +13,19 @@ import PackageJson from '../../package.json';
 import * as path from 'path';
 import { getNpmScope, getNpmScopePrefix } from './npm-scope';
 import { toSnakeCase } from './names';
-import { buildGeneratorInfoList, type GeneratorInfo } from './generators';
 
-export type NxGeneratorInfo = GeneratorInfo;
-export { buildGeneratorInfoList } from './generators';
-
-const GENERATORS = buildGeneratorInfoList(path.resolve(__dirname, '..', '..'));
-
-/**
- * List Nx Plugin for AWS generators
- * @param includeHidden include hidden generators (default false)
- */
-export const listGenerators = (includeHidden = false) =>
-  GENERATORS.filter((g) => includeHidden || !g.hidden);
+export type { NxGeneratorInfo, GeneratorInfo } from './generators';
+export { buildGeneratorInfoList, listGenerators } from './generators';
+import { listGenerators as _listGenerators } from './generators';
 
 /**
  * Return generator information. Call this from a generator method with __filename
  */
-export const getGeneratorInfo = (
-  generatorFileName: string,
-): NxGeneratorInfo => {
+export const getGeneratorInfo = (generatorFileName: string) => {
+  const generators = _listGenerators(true);
   const { dir, name } = path.parse(path.resolve(generatorFileName));
   const resolvedFactoryPath = path.join(dir, name);
-  return GENERATORS.find(
+  return generators.find(
     (generatorInfo) =>
       generatorInfo.resolvedFactoryPath === resolvedFactoryPath,
   );
@@ -75,7 +65,7 @@ export const readProjectConfigurationUnqualified = (
 export const addGeneratorMetadata = (
   tree: Tree,
   projectName: string,
-  info: NxGeneratorInfo,
+  info: { id: string },
   additionalMetadata?: { [key: string]: any },
 ) => {
   const config = readProjectConfigurationUnqualified(tree, projectName);
@@ -105,7 +95,7 @@ export interface ComponentMetadata {
 export const addComponentGeneratorMetadata = (
   tree: Tree,
   projectName: string,
-  info: NxGeneratorInfo,
+  info: { id: string },
   componentPath: string,
   componentName?: string,
   additionalMetadata?: { [key: string]: any },
